fix: don't cast pointer to 32 bit for arithmetic

+++ LR2/En_fileutil.cpp
@@ -819,9 +819,9 @@ void MD5byte(char **iStr, uint len, char *oByte){
 	uVar13 = iVar20 + 0x80 + len;
 	piVar12 = (int *)realloc(*iStr, uVar13);
 	*iStr = (char *)piVar12;
-	memset((undefined *)((int)piVar12 + len), 0, iVar21);
-	piVar16 = (int *)((int)piVar12 + iVar21 + len);
-	*(undefined *)((int)piVar12 + len) = 0x80;
+	memset((undefined *)((uintptr_t)piVar12 + len), 0, iVar21);
+	piVar16 = (int *)((uintptr_t)piVar12 + iVar21 + len);
+	*(undefined *)((uintptr_t)piVar12 + len) = 0x80;
 	*piVar16 = len * 8;
 	piVar16[1] = len >> 0x1d;
 	iVar20 = 0x67452301;
